move home feed view into binders
