+++ apple/OmnivoreKit/Sources/Views/Images/Images.swift
@@ -1,6 +1,6 @@
 import SwiftUI
 
-extension Image {
+public extension Image {
   static var smallOmnivoreLogo: Image { Image("_smallOmnivoreLogo", bundle: .module) }
   static var omnivoreTitleLogo: Image { Image("_omnivoreTitleLogo", bundle: .module) }
   static var readingIllustration: Image { Image("_readingIllustration", bundle: .module) }
